refactor: extract useTipRecipientValidation, useExchangeRateFetcher, useServerSync, usePaymentPolling from Dashboard
Extract 4 useEffect/handler hooks from Dashboard.js:
- useTipRecipientValidation: Blink username validation + debounced check + auto-enable tips
- useExchangeRateFetcher: exchange rate fetching, USD rate for vouchers, tip profile sync
- useServerSync: server preference sync, voucher wallet sync, user change detection
- usePaymentPolling: payment status polling with timeout, NFC Boltcard setup

Dashboard.js: 4,290 -> 3,664 lines (-626 lines)
