+++ lib/hooks/useServerSync.js
@@ -1,233 +1,410 @@
-/**
- * useServerSync - Hook for cross-device data synchronization
- * 
- * Provides functions to sync user data with the server:
- * - Blink Lightning Address wallets
- * - NWC connections
- * - UI preferences
- * 
- * Works alongside existing sync for:
- * - Blink API Key wallets (/api/auth/nostr-blink-account)
- * - Split Payment Profiles (/api/split-profiles)
- */
+import { useEffect, useRef, useCallback } from "react"
+import { getApiUrl } from "../config/api"
+import { initTransactionLabels } from "../../components/TransactionDetail"
 
-import { useState, useCallback, useEffect, useRef } from 'react';
-
-const SYNC_DEBOUNCE_MS = 1000; // Debounce writes to avoid excessive API calls
+// Voucher Wallet storage key (user-scoped for security)
+const VOUCHER_WALLET_OLD_KEY = "blinkpos-voucher-wallet" // Old global key (for cleanup)
+const VOUCHER_WALLET_PREFIX = "blinkpos-voucher-wallet"
 
 /**
- * @typedef {Object} BlinkLnAddressWallet
- * @property {string} id
- * @property {string} label
- * @property {string} username
- * @property {string} lightningAddress
- * @property {string} walletId
- * @property {boolean} isActive
- * @property {string} createdAt
- * @property {string} [lastUsed]
+ * Get user-scoped storage key for voucher wallet
+ * @param {string} userPubkey - User's public key
+ * @returns {string|null} Storage key or null if no user
  */
+export const getVoucherWalletKey = (userPubkey) =>
+  userPubkey ? `${VOUCHER_WALLET_PREFIX}_${userPubkey}` : null
 
 /**
- * @typedef {Object} NWCConnection
- * @property {string} id
- * @property {string} label
- * @property {string} uri - NWC connection string (nostr+walletconnect://...)
- * @property {string[]} [capabilities]
- * @property {boolean} isActive
- * @property {string} createdAt
- * @property {string} [lastUsed]
+ * Clean up old global voucher wallet storage key
+ * This prevents cross-user data leakage from old versions
  */
-
-/**
- * @typedef {Object} Preferences
- * @property {boolean} soundEnabled
- * @property {string} soundTheme
- * @property {boolean} darkMode
- * @property {string} displayCurrency
- * @property {boolean} tipsEnabled
- * @property {number[]} tipPresets
- */
-
-/**
- * @typedef {Object} SyncedData
- * @property {BlinkLnAddressWallet[]} blinkLnAddressWallets
- * @property {NWCConnection[]} nwcConnections
- * @property {Preferences} preferences
- * @property {string} [lastSynced]
- */
-
-export function useServerSync(userPubkey) {
-  const [syncedData, setSyncedData] = useState(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(null);
-  const [lastSynced, setLastSynced] = useState(null);
-  
-  // Debounce timer ref
-  const saveTimerRef = useRef(null);
-  const pendingDataRef = useRef(null);
-
-  /**
-   * Fetch all synced data from server
-   */
-  const fetchSyncedData = useCallback(async () => {
-    if (!userPubkey) {
-      console.log('[useServerSync] No pubkey, skipping fetch');
-      return null;
+const cleanupOldGlobalVoucherWalletStorage = () => {
+  try {
+    if (localStorage.getItem(VOUCHER_WALLET_OLD_KEY)) {
+      console.log("[Dashboard] Removing old global voucher wallet storage (security fix)")
+      localStorage.removeItem(VOUCHER_WALLET_OLD_KEY)
     }
-
-    setLoading(true);
-    setError(null);
-
-    try {
-      console.log('[useServerSync] Fetching synced data for:', userPubkey.substring(0, 8) + '...');
-      
-      const response = await fetch(`/api/user/sync?pubkey=${userPubkey}`);
-      
-      if (!response.ok) {
-        throw new Error(`Server error: ${response.status}`);
-      }
-
-      const data = await response.json();
-      console.log('[useServerSync] Fetched data:', {
-        lnAddressWallets: data.blinkLnAddressWallets?.length || 0,
-        nwcConnections: data.nwcConnections?.length || 0,
-        hasPreferences: !!data.preferences
-      });
-
-      setSyncedData(data);
-      setLastSynced(data.lastSynced);
-      return data;
-    } catch (err) {
-      console.error('[useServerSync] Fetch error:', err);
-      setError(err.message);
-      return null;
-    } finally {
-      setLoading(false);
-    }
-  }, [userPubkey]);
-
-  /**
-   * Save all synced data to server (debounced)
-   */
-  const saveSyncedData = useCallback(async (data, immediate = false) => {
-    if (!userPubkey) {
-      console.log('[useServerSync] No pubkey, skipping save');
-      return false;
-    }
-
-    // Store pending data
-    pendingDataRef.current = data;
-
-    // Clear existing timer
-    if (saveTimerRef.current) {
-      clearTimeout(saveTimerRef.current);
-    }
-
-    // If immediate, save now
-    if (immediate) {
-      return performSave(userPubkey, data);
-    }
-
-    // Otherwise debounce
-    return new Promise((resolve) => {
-      saveTimerRef.current = setTimeout(async () => {
-        const result = await performSave(userPubkey, pendingDataRef.current);
-        resolve(result);
-      }, SYNC_DEBOUNCE_MS);
-    });
-  }, [userPubkey]);
-
-  /**
-   * Perform the actual save operation
-   */
-  const performSave = async (pubkey, data) => {
-    try {
-      console.log('[useServerSync] Saving data for:', pubkey.substring(0, 8) + '...');
-      
-      const response = await fetch('/api/user/sync', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          pubkey,
-          ...data
-        })
-      });
-
-      if (!response.ok) {
-        throw new Error(`Server error: ${response.status}`);
-      }
-
-      const result = await response.json();
-      setLastSynced(result.lastSynced);
-      console.log('[useServerSync] ✓ Data saved successfully');
-      return true;
-    } catch (err) {
-      console.error('[useServerSync] Save error:', err);
-      setError(err.message);
-      return false;
-    }
-  };
-
-  /**
-   * Update a specific field (PATCH request)
-   */
-  const updateField = useCallback(async (field, data) => {
-    if (!userPubkey) {
-      console.log('[useServerSync] No pubkey, skipping update');
-      return false;
-    }
-
-    try {
-      console.log(`[useServerSync] Updating field '${field}'`);
-      
-      const response = await fetch('/api/user/sync', {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          pubkey: userPubkey,
-          field,
-          data
-        })
-      });
-
-      if (!response.ok) {
-        throw new Error(`Server error: ${response.status}`);
-      }
-
-      const result = await response.json();
-      setLastSynced(result.lastSynced);
-      
-      // Update local state
-      setSyncedData(prev => prev ? { ...prev, [field]: data } : null);
-      
-      console.log(`[useServerSync] ✓ Field '${field}' updated`);
-      return true;
-    } catch (err) {
-      console.error('[useServerSync] Update error:', err);
-      setError(err.message);
-      return false;
-    }
-  }, [userPubkey]);
-
-  // Cleanup on unmount
-  useEffect(() => {
-    return () => {
-      if (saveTimerRef.current) {
-        clearTimeout(saveTimerRef.current);
-      }
-    };
-  }, []);
-
-  return {
-    syncedData,
-    loading,
-    error,
-    lastSynced,
-    fetchSyncedData,
-    saveSyncedData,
-    updateField
-  };
+  } catch (err) {
+    console.error("[Dashboard] Failed to cleanup old voucher wallet storage:", err)
+  }
 }
 
-export default useServerSync;
+/**
+ * Hook for syncing preferences and voucher wallet data to/from the server.
+ *
+ * Extracted from Dashboard.js — contains:
+ * - syncPreferencesToServer (debounced 2s)
+ * - syncVoucherWalletToServer
+ * - Fetch preferences from server on login
+ * - Sync preferences on change
+ * - Voucher wallet server sync and migration
+ * - User change detection (clear voucher wallet)
+ * - Cleanup timer on unmount
+ *
+ * @param {Object} params - All required state and setters
+ * @returns {Object} { syncVoucherWalletToServer, getVoucherWalletKey }
+ */
+export function useServerSync({
+  publicKey,
+  soundEnabled,
+  setSoundEnabled,
+  soundTheme,
+  setSoundTheme,
+  tipsEnabled,
+  setTipsEnabled,
+  tipPresets,
+  setTipPresets,
+  displayCurrency,
+  setDisplayCurrency,
+  numberFormat,
+  setNumberFormat,
+  numpadLayout,
+  setNumpadLayout,
+  voucherCurrencyMode,
+  setVoucherCurrencyMode,
+  voucherExpiry,
+  setVoucherExpiry,
+  voucherWallet,
+  setVoucherWallet,
+  setVoucherWalletBalance,
+  setVoucherWalletUsdBalance,
+}) {
+  const serverSyncTimerRef = useRef(null)
+  const lastSyncedPrefsRef = useRef(null)
 
+  // Sync preferences to server (debounced)
+  const syncPreferencesToServer = useCallback(
+    async (prefs) => {
+      if (!publicKey) return
+
+      // Clear existing timer
+      if (serverSyncTimerRef.current) {
+        clearTimeout(serverSyncTimerRef.current)
+      }
+
+      // Debounce the sync (2 seconds)
+      serverSyncTimerRef.current = setTimeout(async () => {
+        try {
+          console.log("[Dashboard] Syncing preferences to server...")
+
+          const response = await fetch("/api/user/sync", {
+            method: "PATCH",
+            headers: { "Content-Type": "application/json" },
+            credentials: "include", // Include session cookie for auth
+            body: JSON.stringify({
+              pubkey: publicKey,
+              field: "preferences",
+              data: prefs,
+            }),
+          })
+
+          if (response.ok) {
+            console.log("[Dashboard] ✓ Preferences synced to server")
+            lastSyncedPrefsRef.current = JSON.stringify(prefs)
+          }
+        } catch (err) {
+          console.error("[Dashboard] Server sync error:", err)
+        }
+      }, 2000)
+    },
+    [publicKey],
+  )
+
+  // Sync voucher wallet to server
+  const syncVoucherWalletToServer = useCallback(
+    async (walletData) => {
+      if (!publicKey) return
+
+      try {
+        console.log("[Dashboard] Syncing voucher wallet to server...")
+        const response = await fetch("/api/user/sync", {
+          method: "PATCH",
+          headers: { "Content-Type": "application/json" },
+          credentials: "include", // Include session cookie for auth
+          body: JSON.stringify({
+            pubkey: publicKey,
+            field: "voucherWallet",
+            data: walletData,
+          }),
+        })
+
+        if (response.ok) {
+          console.log("[Dashboard] ✓ Voucher wallet synced to server")
+        }
+      } catch (err) {
+        console.error("[Dashboard] Failed to sync voucher wallet:", err)
+      }
+    },
+    [publicKey],
+  )
+
+  // Fetch preferences from server on login
+  useEffect(() => {
+    if (!publicKey) return
+
+    const fetchServerPreferences = async () => {
+      try {
+        console.log("[Dashboard] Fetching preferences from server...")
+        const response = await fetch(`/api/user/sync?pubkey=${publicKey}`, {
+          credentials: "include", // Include session cookie for auth
+        })
+
+        if (!response.ok) return
+
+        const data = await response.json()
+        const serverPrefs = data.preferences
+
+        if (serverPrefs) {
+          console.log("[Dashboard] Loaded preferences from server")
+
+          // Apply server preferences to local state
+          if (serverPrefs.soundEnabled !== undefined) {
+            setSoundEnabled(serverPrefs.soundEnabled)
+            localStorage.setItem("soundEnabled", JSON.stringify(serverPrefs.soundEnabled))
+          }
+          if (serverPrefs.soundTheme) {
+            setSoundTheme(serverPrefs.soundTheme)
+            localStorage.setItem("soundTheme", serverPrefs.soundTheme)
+          }
+          if (serverPrefs.tipsEnabled !== undefined) {
+            setTipsEnabled(serverPrefs.tipsEnabled)
+            localStorage.setItem(
+              "blinkpos-tips-enabled",
+              serverPrefs.tipsEnabled.toString(),
+            )
+          }
+          if (serverPrefs.tipPresets) {
+            setTipPresets(serverPrefs.tipPresets)
+            localStorage.setItem(
+              "blinkpos-tip-presets",
+              JSON.stringify(serverPrefs.tipPresets),
+            )
+          }
+          if (serverPrefs.displayCurrency) {
+            setDisplayCurrency(serverPrefs.displayCurrency)
+          }
+          if (serverPrefs.numberFormat) {
+            setNumberFormat(serverPrefs.numberFormat)
+            localStorage.setItem("blinkpos-number-format", serverPrefs.numberFormat)
+          }
+          if (serverPrefs.numpadLayout) {
+            setNumpadLayout(serverPrefs.numpadLayout)
+            localStorage.setItem("blinkpos-numpad-layout", serverPrefs.numpadLayout)
+          }
+          if (serverPrefs.voucherCurrencyMode) {
+            setVoucherCurrencyMode(serverPrefs.voucherCurrencyMode)
+            localStorage.setItem(
+              "blinkpos-voucher-currency-mode",
+              serverPrefs.voucherCurrencyMode,
+            )
+          }
+          // Handle voucherExpiry with migration from old default '7d' to new default '24h'
+          if (serverPrefs.voucherExpiry) {
+            // Migration: if server has '7d' or legacy values, migrate to '24h'
+            const migratedExpiry =
+              serverPrefs.voucherExpiry === "7d" ||
+              serverPrefs.voucherExpiry === "15m" ||
+              serverPrefs.voucherExpiry === "1h"
+                ? "24h"
+                : serverPrefs.voucherExpiry
+            setVoucherExpiry(migratedExpiry)
+            localStorage.setItem("blinkpos-voucher-expiry", migratedExpiry)
+          }
+
+          lastSyncedPrefsRef.current = JSON.stringify(serverPrefs)
+        } else {
+          // No server preferences - sync current local to server
+          const currentPrefs = {
+            soundEnabled,
+            soundTheme,
+            tipsEnabled,
+            tipPresets,
+            displayCurrency,
+            numberFormat,
+            numpadLayout,
+            voucherCurrencyMode,
+            voucherExpiry,
+          }
+          syncPreferencesToServer(currentPrefs)
+        }
+
+        // Initialize transaction labels from server
+        await initTransactionLabels()
+        console.log("[Dashboard] Transaction labels synced from server")
+
+        // Clean up old global voucher wallet key (security fix for cross-profile leakage)
+        cleanupOldGlobalVoucherWalletStorage()
+
+        // Sync voucher wallet from server (using user-scoped localStorage key)
+        const voucherWalletStorageKey = getVoucherWalletKey(publicKey)
+        console.log(
+          "[Dashboard] voucherWallet from server:",
+          data.voucherWallet
+            ? {
+                label: data.voucherWallet.label,
+                hasApiKey: !!data.voucherWallet.apiKey,
+                apiKeyLength: data.voucherWallet.apiKey?.length || 0,
+                apiKeyType: typeof data.voucherWallet.apiKey,
+              }
+            : "null/undefined",
+        )
+
+        if (data.voucherWallet && data.voucherWallet.apiKey) {
+          console.log(
+            "[Dashboard] Loaded voucher wallet from server:",
+            data.voucherWallet.label,
+          )
+          setVoucherWallet(data.voucherWallet)
+          if (voucherWalletStorageKey) {
+            localStorage.setItem(
+              voucherWalletStorageKey,
+              JSON.stringify(data.voucherWallet),
+            )
+          }
+        } else if (!data.voucherWallet) {
+          // Check if we have local voucher wallet (user-scoped) to sync to server
+          if (voucherWalletStorageKey) {
+            const localVoucherWallet = localStorage.getItem(voucherWalletStorageKey)
+            if (localVoucherWallet) {
+              const parsed = JSON.parse(localVoucherWallet)
+              console.log("[Dashboard] Syncing local voucher wallet to server")
+              syncVoucherWalletToServer(parsed)
+              setVoucherWallet(parsed)
+            }
+          }
+        }
+      } catch (err) {
+        console.error("[Dashboard] Failed to fetch server preferences:", err)
+      }
+    }
+
+    fetchServerPreferences()
+  }, [publicKey]) // eslint-disable-line react-hooks/exhaustive-deps
+
+  // Track previous user to detect user changes
+  const prevUserRef = useRef(publicKey)
+
+  // Clear voucher wallet state when user changes (logout/switch user)
+  // This prevents cross-user data leakage
+  useEffect(() => {
+    const prevUser = prevUserRef.current
+
+    // User has changed (including logout where publicKey becomes null/undefined)
+    if (prevUser !== publicKey) {
+      console.log("[Dashboard] User changed, clearing voucher wallet state")
+      setVoucherWallet(null)
+      setVoucherWalletBalance(null)
+      setVoucherWalletUsdBalance(null)
+      prevUserRef.current = publicKey
+    }
+  }, [publicKey])
+
+  // Migration: Fetch missing username for voucher wallet (for wallets created before username was added)
+  useEffect(() => {
+    const migrateVoucherWalletUsername = async () => {
+      if (!voucherWallet || !voucherWallet.apiKey || voucherWallet.username) {
+        return // No wallet, no API key, or already has username
+      }
+
+      console.log("[Dashboard] Migrating voucher wallet: fetching username...")
+
+      try {
+        const response = await fetch(getApiUrl(), {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+            "X-API-KEY": voucherWallet.apiKey,
+          },
+          body: JSON.stringify({
+            query: "{ me { username } }",
+          }),
+        })
+
+        if (!response.ok) {
+          console.warn(
+            "[Dashboard] Failed to fetch username for voucher wallet migration",
+          )
+          return
+        }
+
+        const data = await response.json()
+        const username = data.data?.me?.username
+
+        if (username) {
+          console.log("[Dashboard] ✓ Voucher wallet username fetched:", username)
+
+          // Update wallet data with username
+          const updatedWallet = { ...voucherWallet, username }
+          setVoucherWallet(updatedWallet)
+
+          // Save to localStorage (user-scoped)
+          if (typeof window !== "undefined" && publicKey) {
+            const storageKey = getVoucherWalletKey(publicKey)
+            if (storageKey) {
+              localStorage.setItem(storageKey, JSON.stringify(updatedWallet))
+            }
+          }
+
+          // Sync to server
+          syncVoucherWalletToServer(updatedWallet)
+        }
+      } catch (err) {
+        console.error("[Dashboard] Failed to migrate voucher wallet username:", err)
+      }
+    }
+
+    migrateVoucherWalletUsername()
+  }, [voucherWallet?.apiKey]) // Only run when voucherWallet.apiKey changes (initial load)
+
+  // Sync preferences to server when they change
+  useEffect(() => {
+    if (!publicKey) return
+
+    const currentPrefs = {
+      soundEnabled,
+      soundTheme,
+      tipsEnabled,
+      tipPresets,
+      displayCurrency,
+      numberFormat,
+      numpadLayout,
+      voucherCurrencyMode,
+      voucherExpiry,
+    }
+
+    const currentPrefsStr = JSON.stringify(currentPrefs)
+
+    // Only sync if preferences actually changed (avoid initial sync loop)
+    if (lastSyncedPrefsRef.current && lastSyncedPrefsRef.current !== currentPrefsStr) {
+      syncPreferencesToServer(currentPrefs)
+    }
+  }, [
+    publicKey,
+    soundEnabled,
+    soundTheme,
+    tipsEnabled,
+    tipPresets,
+    displayCurrency,
+    numberFormat,
+    numpadLayout,
+    voucherCurrencyMode,
+    voucherExpiry,
+    syncPreferencesToServer,
+  ])
+
+  // Cleanup server sync timer on unmount
+  useEffect(() => {
+    return () => {
+      if (serverSyncTimerRef.current) {
+        clearTimeout(serverSyncTimerRef.current)
+      }
+    }
+  }, [])
+
+  return {
+    syncVoucherWalletToServer,
+    getVoucherWalletKey,
+  }
+}
+
+export default useServerSync
